+++ lib/caravela/gen/svelte_form.ex
@@ -161,7 +161,7 @@ defmodule Caravela.Gen.SvelteForm do
     <input
         type="checkbox"
         checked={#{row_var}.#{name} ?? false}
-        on:change={(e) => handleChange('#{name}', e.currentTarget.checked)}
+        onchange={(e) => handleChange('#{name}', e.currentTarget.checked)}
       />\
     """
   end
@@ -170,7 +170,7 @@ defmodule Caravela.Gen.SvelteForm do
     """
     <textarea
         value={#{row_var}.#{name} ?? ''}
-        on:input={(e) => handleChange('#{name}', e.currentTarget.value)}
+        oninput={(e) => handleChange('#{name}', e.currentTarget.value)}
       />\
     """
   end
@@ -181,7 +181,7 @@ defmodule Caravela.Gen.SvelteForm do
     <input
         type="number"
         value={#{row_var}.#{name} ?? ''}
-        on:input={(e) => handleChange('#{name}', e.currentTarget.value)}
+        oninput={(e) => handleChange('#{name}', e.currentTarget.value)}
       />\
     """
   end
@@ -191,7 +191,7 @@ defmodule Caravela.Gen.SvelteForm do
     <input
         type="date"
         value={#{row_var}.#{name} ?? ''}
-        on:input={(e) => handleChange('#{name}', e.currentTarget.value)}
+        oninput={(e) => handleChange('#{name}', e.currentTarget.value)}
       />\
     """
   end
@@ -201,7 +201,7 @@ defmodule Caravela.Gen.SvelteForm do
     <input
         type="datetime-local"
         value={#{row_var}.#{name} ?? ''}
-        on:input={(e) => handleChange('#{name}', e.currentTarget.value)}
+        oninput={(e) => handleChange('#{name}', e.currentTarget.value)}
       />\
     """
   end
@@ -211,7 +211,7 @@ defmodule Caravela.Gen.SvelteForm do
     <input
         type="text"
         value={#{row_var}.#{name} ?? ''}
-        on:input={(e) => handleChange('#{name}', e.currentTarget.value)}
+        oninput={(e) => handleChange('#{name}', e.currentTarget.value)}
       />\
     """
   end